mammo_radiomics/requirements.txt
pyradiomics
SimpleITK
pandas
numpy